nfc nfca: rework poller
