+++ applications/main/nfc/nfc_app.c
@@ -46,7 +46,6 @@ NfcApp* nfc_app_alloc() {
         instance->view_dispatcher, nfc_back_event_callback);
 
     instance->nfc = nfc_alloc();
-    instance->nfca_poller = nfca_poller_alloc(instance->nfc);
     instance->nfca_listener = nfca_listener_alloc(instance->nfc);
     instance->mf_ul_listener = mf_ultralight_listener_alloc(instance->nfca_listener);
 
@@ -156,7 +155,6 @@ void nfc_app_free(NfcApp* instance) {
 
     mf_ultralight_listener_free(instance->mf_ul_listener);
     nfca_listener_free(instance->nfca_listener);
-    nfca_poller_free(instance->nfca_poller);
     nfc_free(instance->nfc);
     nfc_scanner_free(instance->scanner);
