+++ lib/nfc/protocols/nfca/nfca_poller.h
@@ -22,26 +22,10 @@ typedef struct {
 
 typedef struct {
     NfcaPollerEventType type;
-    NfcaPollerEventData data;
+    NfcaPollerEventData* data;
 } NfcaPollerEvent;
 
-typedef enum {
-    NfcaPollerCommandContinue = NfcCommandContinue,
-    NfcaPollerCommandReset = NfcCommandReset,
-    NfcaPollerCommandStop = NfcCommandStop,
-} NfcaPollerCommand;
-
-typedef NfcaPollerCommand (*NfcaPollerEventCallback)(NfcaPollerEvent event, void* context);
-
-NfcaPoller* nfca_poller_alloc(Nfc* nfc);
-
-void nfca_poller_free(NfcaPoller* instance);
-
 const NfcaData* nfca_poller_get_data(NfcaPoller* instance);
-
-NfcaError nfca_poller_start(NfcaPoller* instance, NfcaPollerEventCallback callback, void* context);
-
-NfcaError nfca_poller_stop(NfcaPoller* instance);
 
 #ifdef __cplusplus
 }
