+++ lib/nfc/protocols/nfca/nfca_poller_i.c
@@ -1,7 +1,5 @@
 #include "nfca_poller_i.h"
 
-#include <nfc/protocols/nfc_poller_base.h>
-
 #include <furi.h>
 
 #define TAG "NFCA"
@@ -347,123 +345,3 @@ NfcaError nfca_poller_send_standart_frame(
 
     return ret;
 }
-
-static NfcaPoller* nfca_poller_alloc_new(Nfc* nfc) {
-    furi_assert(nfc);
-
-    NfcaPoller* instance = malloc(sizeof(NfcaPoller));
-    instance->nfc = nfc;
-    instance->tx_buffer = bit_buffer_alloc(NFCA_POLLER_MAX_BUFFER_SIZE);
-    instance->rx_buffer = bit_buffer_alloc(NFCA_POLLER_MAX_BUFFER_SIZE);
-
-    nfc_config(instance->nfc, NfcModeNfcaPoller);
-    nfc_set_guard_time_us(instance->nfc, NFCA_GUARD_TIME_US);
-    nfc_set_fdt_poll_fc(instance->nfc, NFCA_FDT_POLL_FC);
-    nfc_set_fdt_poll_poll_us(instance->nfc, NFCA_POLL_POLL_MIN_US);
-    instance->data = nfca_alloc();
-    instance->event = malloc(sizeof(NfcaPollerEvent));
-
-    return instance;
-}
-
-static void nfca_poller_free_new(NfcaPoller* nfca_poller) {
-    furi_assert(nfca_poller);
-
-    NfcaPoller* instance = nfca_poller;
-    furi_assert(instance->tx_buffer);
-    furi_assert(instance->rx_buffer);
-    furi_assert(instance->data);
-    furi_assert(instance->event);
-
-    bit_buffer_free(instance->tx_buffer);
-    bit_buffer_free(instance->rx_buffer);
-    nfca_free(instance->data);
-    free(instance->event);
-    free(instance);
-}
-
-static void
-    nfca_poller_set_callback(NfcaPoller* poller, NfcPollerCallback callback, void* context) {
-    furi_assert(poller);
-    furi_assert(callback);
-
-    NfcaPoller* instance = poller;
-    instance->callback_new = callback;
-    instance->context_new = context;
-}
-
-static NfcCommand nfca_poller_run(NfcPollerEvent event, void* context) {
-    furi_assert(context);
-
-    NfcaPoller* instance = context;
-    NfcEvent* nfc_event = event.data;
-    NfcCommand command = NfcCommandContinue;
-    NfcPollerEvent nfca_poller_event = {
-        .protocol_type = NfcProtocolTypeIso14443_3a,
-        .poller = instance,
-    };
-
-    if(nfc_event->type == NfcEventTypePollerReady) {
-        if(instance->state != NfcaPollerStateActivated) {
-            NfcaData data = {};
-            NfcaError error = nfca_poller_async_activate(instance, &data);
-            if(error == NfcaErrorNone) {
-                instance->event->type = NfcaPollerEventTypeReady;
-                instance->event->data.error = error;
-                instance->state = NfcaPollerStateActivated;
-                nfca_poller_event.data = instance->event;
-                command = instance->callback_new(nfca_poller_event, instance->context_new);
-            } else {
-                instance->event->type = NfcaPollerEventTypeError;
-                instance->event->data.error = error;
-                nfca_poller_event.data = instance->event;
-                command = instance->callback_new(nfca_poller_event, instance->context_new);
-                // Add delay to switch context
-                furi_delay_ms(100);
-            }
-        } else {
-            instance->event->type = NfcaPollerEventTypeReady;
-            instance->event->data.error = NfcaErrorNone;
-            nfca_poller_event.data = instance->event;
-            command = instance->callback_new(nfca_poller_event, instance->context_new);
-        }
-    }
-
-    return command;
-}
-
-static bool nfca_poller_detect(NfcPollerEvent event, void* context) {
-    furi_assert(context);
-    furi_assert(event.data);
-    furi_assert(event.poller);
-
-    bool protocol_detected = false;
-    NfcaPoller* instance = context;
-    NfcEvent* nfc_event = event.data;
-    furi_assert(instance->state == NfcaPollerStateIdle);
-
-    if(nfc_event->type == NfcEventTypePollerReady) {
-        NfcaError error = nfca_poller_async_activate(instance, NULL);
-        protocol_detected = (error == NfcaErrorNone);
-    }
-
-    return protocol_detected;
-}
-
-static const NfcProtocolData* nfca_poller_get_data_new(const NfcaPoller* nfca_poller) {
-    furi_assert(nfca_poller);
-
-    const NfcaPoller* instance = nfca_poller;
-    furi_assert(instance->data);
-
-    return instance->data;
-}
-
-const NfcPollerBase nfc_poller_iso14443_3a = {
-    .alloc = (NfcPollerAlloc)nfca_poller_alloc_new,
-    .free = (NfcPollerFree)nfca_poller_free_new,
-    .set_callback = (NfcPollerSetCallback)nfca_poller_set_callback,
-    .run = (NfcPollerRun)nfca_poller_run,
-    .detect = (NfcPollerDetect)nfca_poller_detect,
-    .get_data = (NfcPollerGetData)nfca_poller_get_data_new,
-};
